Split tests, add test for hessian and bestfit
